requirements.txt
fastapi
uvicorn[standard]
gunicorn
psycopg2-binary
redis
pydantic
python-dotenv